Added auto load cog using for loop and added unban command
