+++ bot.py
@@ -15,19 +15,15 @@ from contextlib import redirect_stdout
 from discord.ext import commands
 import json
 bot = commands.Bot(command_prefix=commands.when_mentioned_or('*'),description="The revamped dat banana bot made by dat banana boi#1982.\n\nHelp Commands",owner_id=277981712989028353)
-bot.remove_command("help")
-bot.load_extension("cogs.math")
-bot.load_extension("cogs.mod")
-bot.load_extension("cogs.utility")
-bot.load_extension("cogs.fun")
-bot.load_extension("cogs.info")
-bot.load_extension("cogs.developer")
-bot.load_extension("cogs.cr")
-bot.load_extension("cogs.help")
-bot.load_extension("cogs.coc")
-#bot.load_extension("cogs.lol")
 
 
+for filename in os.listdir('./cogs'):
+    if filename.endswith('.py') and not filename.startswith('_'):
+        bot.load_extension(f'cogs.{filename[:-3]}')
+        print('Successfully loaded {filename[:-3]}')
+
+    else:
+        print('Unable to load {filename[:-3]}')
 
 def cleanup_code(content):
     # remove ```py\n```
